#106: Improved object-orientation and replaced global exa access with injection.
